# lcm-engine/lcm_engine/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: |
    The Life-Cycle Management Engine API specification for the Gaia-X project.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: LCM Engine API
  version: 2.3.0
servers:
- description: Testing LCM Engine server 1
  url: https://demo.gxfs.xlab.si
- description: Testing LCM Engine server 2
  url: https://orchestrator.gxfs.xlab.si
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /workspace:
    get:
      operationId: get_workspaces
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Workspace'
                type: array
          description: A list of workspaces
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
      summary: List available workspaces
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
    post:
      operationId: create_workspace
      parameters:
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/WorkspaceBody'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: The created Workspace
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
      summary: Create a new workspace
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
  /workspace/{workspaceId}:
    delete:
      operationId: delete_workspace
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Workspace deleted
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Delete a workspace
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
    get:
      operationId: describe_workspace
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: A workspace
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Describe a workspace
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
    patch:
      operationId: update_workspace
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/WorkspaceBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: The updated workspace
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Update a workspace
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
    put:
      operationId: replace_workspace
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/WorkspaceBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: The updated workspace
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Replace a workspace
      tags:
      - workspace
      x-openapi-router-controller: lcm_engine.controllers.workspace_controller
...
components:
...
  requestBodies:
...
    WorkspaceBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Workspace'
      description: Workspace specification object
      required: true
...
  schemas:
...
    Workspace:
      additionalProperties: false
      description: A workspace
      example:
        projects:
        - 0
        - 4
        - 6
        isOwner: false
        name: my workspace
        id: 5
        secrets:
        - 2
        - 6
        - 4
      properties:
        id:
          example: 5
          format: int64
          readOnly: true
          title: id
          type: integer
        name:
          example: my workspace
          title: name
          type: string
        secrets:
          example:
          - 2
          - 6
          - 4
          format: int64
          items:
            type: integer
          readOnly: true
          title: secrets
          type: array
        projects:
          example:
          - 0
          - 4
          - 6
          format: int64
          items:
            type: integer
          readOnly: true
          title: projects
          type: array
        isOwner:
          example: false
          readOnly: true
          title: isOwner
          type: boolean
      required:
      - id
      - isOwner
      - name
      - projects
      - secrets
      title: Workspace
      type: object
...
    Error:
      additionalProperties: false
      description: An error
      properties:
        msg:
          title: msg
          type: string
      required:
      - msg
      title: Error
      type: object
...